requirements.txt
pymysql==1.1.0
SQLAlchemy==2.0.27
apscheduler==3.10.4
jinja2==3.1.3

python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
alembic==1.13.1
pytest==8.0.0
black==24.1.1
flake8==7.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9